+++ lib/phoenix_feathers.ex
@@ -2,17 +2,4 @@ defmodule PhoenixFeathers do
   @moduledoc """
   Documentation for PhoenixFeathers.
   """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> PhoenixFeathers.hello()
-      :world
-
-  """
-  def hello do
-    :world
-  end
 end