+++ lib/phoenix_feathers/live_component.ex
@@ -1,0 +1,8 @@
+defmodule PhoenixFeathers.LiveComponent do
+  defmacro __using__(_opts) do
+    quote do
+      use Phoenix.LiveComponent
+      use Phoenix.HTML
+    end
+  end
+end