+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.1.0]
+## [0.1.0-alpha]
 ### Added
-* Basic functionality.
+* Basic button component
+* Basic card component
+* Component tests